python: Generalize subprocess cleanup, and use TERM instead of HUP
Trying to terminate our children with SIGHUP was slightly odd anyway (as
the canonical signal is TERM). We did that because the C bridge does.

However, this does not work with the Python bridge on Debian/Ubuntu: its
PAM stack does not reset the signal mask, and the Python bridge does not
configure the signals or do any cleanups on SIGTERM either. As
cockpit-session ignores SIGHUP, we ended up with ignoring SIGHUP in the
user and priv bridges and all processes spawned by it.

While that unintended inheritance of SIGHUP should be fixed, let's also
try to avoid it and send SIGTERM to the bridge children on exit.

Also, this really applies to all processes that the bridge spawns: We
don't want to leak any process, not just the "spawn" channel ones in
particular. Thus lift the PDEATHSIG up to SubprocessTransport.
